Move test from progress to loops
